refactor(transit): group transit resolvers
